restructures desktop directory to differentiate between runner and user components
