+++ cmd/launcher/desktop.go
@@ -15,10 +15,10 @@ import (
 	"github.com/go-kit/kit/log/level"
 	"github.com/kolide/kit/logutil"
 	"github.com/kolide/kit/ulid"
-	"github.com/kolide/launcher/ee/desktop/menu"
-	"github.com/kolide/launcher/ee/desktop/notify"
 	runnerserver "github.com/kolide/launcher/ee/desktop/runner/server"
-	userserver "github.com/kolide/launcher/ee/desktop/server"
+	"github.com/kolide/launcher/ee/desktop/user/menu"
+	"github.com/kolide/launcher/ee/desktop/user/notify"
+	userserver "github.com/kolide/launcher/ee/desktop/user/server"
 	"github.com/kolide/launcher/pkg/agent"
 	"github.com/kolide/launcher/pkg/authedclient"
 	"github.com/kolide/systray"
