+++ ee/desktop/runner/runner.go
@@ -22,10 +22,10 @@ import (
 	"github.com/kolide/kit/ulid"
 	"github.com/kolide/kit/version"
 	"github.com/kolide/launcher/ee/consoleuser"
-	"github.com/kolide/launcher/ee/desktop/client"
-	"github.com/kolide/launcher/ee/desktop/menu"
-	"github.com/kolide/launcher/ee/desktop/notify"
 	runnerserver "github.com/kolide/launcher/ee/desktop/runner/server"
+	"github.com/kolide/launcher/ee/desktop/user/client"
+	"github.com/kolide/launcher/ee/desktop/user/menu"
+	"github.com/kolide/launcher/ee/desktop/user/notify"
 	"github.com/kolide/launcher/ee/ui/assets"
 	"github.com/kolide/launcher/pkg/agent"
 	"github.com/kolide/launcher/pkg/agent/flags/keys"
