+++ ee/desktop/user/client/dialer_windows.go
