+++ ee/desktop/user/menu/action_open_url.go
