+++ ee/desktop/user/menu/action_open_url_darwin.go
