+++ ee/desktop/user/menu/action_test.go
