+++ ee/desktop/user/menu/initial_menu.json
