+++ ee/desktop/user/menu/menu_icon.go
