+++ ee/desktop/user/menu/menu_icon_test.go
