+++ ee/desktop/user/menu/menu_parse.go
