+++ ee/desktop/user/menu/menu_parse_test.go
