+++ ee/desktop/user/menu/menu_systray.go
