+++ ee/desktop/user/menu/menu_template.go
