+++ ee/desktop/user/menu/menu_template_test.go
